worker/requirements.txt
fvcore==0.1.5.post20221221
layoutparser==0.3.4
pytesseract==0.3.13
reportlab==4.2.5
shapely==2.0.6
openai==1.60.1
celery==5.4.0
redis==5.2.1
pdf2image==1.17.0
numpy==1.24.4
pydantic-settings
